// src/app/api/model/cu-loan-product-model.ts
/**
 * CrediFlow.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CULoanProductModel { 
    loanProductId?: string | null;
    storeId?: string | null;
    productCode: string;
    productName: string;
    description?: string | null;
    minPrincipalAmount?: number;
    maxPrincipalAmount?: number;
    minTermMonths?: number;
    maxTermMonths?: number;
    interestRateMonthly?: number;
    qlkvRateMonthly?: number;
    qltsRateMonthly?: number;
    fixedMonthlyFeeAmount?: number;
    defaultFileFeeAmount?: number;
    defaultInsuranceRate?: number;
    isActive?: boolean;
}